automated-safety-monitoring/requirements.txt
ultralytics>=8.0.0
opencv-python
numpy<2
pandas
torch==2.1.0
torchvision==0.16.0
matplotlib
flask
flask-cors
face_recognition
dlib
fastapi
uvicorn
python-multipart
pillow
openpyxl
psycopg2-binary
sqlalchemy